requirements.txt
fastapi
uvicorn[standard]
httpx
python-pptx
python-multipart